Move conversion of sample_idx=0 to current time into net connector.
This was previously in the DigitalRFSink, but that means other operators
relying on the metadata's sample index would not get that subsitution
the converts an initial sample_idx of 0 to the sample index
corresponding to the current time. That really belongs where the spoofed
header is generated, since that's where that conversion is relevant,
even if it means that the time used is a little earlier than when the
packets might actually be sampled.