rf_array_net_connector: Fix buffer tracking, add some logging
Incrementing the buffer tracking counter on the host had edge case
problems that would cause the counter to be *ahead* of the samples,
which would result in samples not being placed into that buffer and thus
dropped samples. Since the place_packet_data kernel can keep the
tracker updated on its own based on the incoming packets, this removes
the host-side increment and fixes the bug.