basic_network: Clear UDP packet buffer on first compute() call
A variable amount of time can pass between initialize() and the first
compute() which could result in the packet buffer filling up. Since we
don't want to start behind, and typical use would prefer to start with a
continuous string of packets instead of ones with a possible gap after
one full buffer is captured, clear the UDP packet buffer on the first
compute() call and process new packets from that point forward.